requirements.txt
Django>=4.2.0,<5.0.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
msal>=1.27.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
